requirements.txt

langchain>=0.1.0
langchain-community>=0.0.16
langchain-openai>=0.0.5
langchain-core>=0.1.15
langchain-qdrant>=0.0.1
langgraph>=0.0.21
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.3
qdrant-client>=1.7.0
langsmith>=0.0.78
openai>=1.6.0
pypdf>=4.0.0
